SWITCHYARD-2510 Enable @PreDestroy for SY managed CDI Beans (Bean-Service, Transformer, Validator)
